requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
numpy==2.4.2
openai==2.21.0
pandas==3.0.1
pydantic==2.12.5
pydantic-core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyyaml==6.0.3
six==1.17.0
sniffio==1.3.1
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
